+++ src/page/Login/Login7.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { useLocation, useNavigate } from "react-router-dom"; // useNavigate 추가
 import { IoChevronBack } from "react-icons/io5";
-import '../LoginPage/SignIn.css';  
+import './Login.css';  
 
 // 생년월일 입력
 const Login7 = () => {
@@ -38,13 +38,16 @@ const Login7 = () => {
         <button className="back-button" onClick={handleBack}>
           <IoChevronBack />  
         </button>
+
         <div className="progress-bar">
           <div className="progress" style={{ width: "80%" }}></div>
         </div>
       </header>
+
+      <div className="container">
       <h2 className="login-heading">춘식이님의 생일을 알려주세요</h2>
       <form onSubmit={handleSubmit}>
-        <div className = "year-month-day-container">
+        <div className="year-month-day-container">
           <input
             type="text"
             maxLength="4"
@@ -70,11 +73,12 @@ const Login7 = () => {
             className="day-input"
           />
         </div>
-
+    
         <button type="submit" className="bottom-Button">
           다음
         </button>
       </form>
+      </div>
     </div>
   );
 };
